Add support for generated source files
Generated files are available under different paths than source files in
the execroot. E.g.
```
./bazel-out/k8-fastbuild/bin/strip-prefix
```
instead of
```
./strip-prefix
```

However, we cannot just always add that path to the source directories.
Otherwise, the Elm compiler will complain about a missing directory if
there are no generated sources to a target.

This change avoids that issue by using the `File.root` attribute on each
source file to determine if an additional path component is required.